Move changes to its own submodule
This way the drivers can use it
