+++ changes/changes.go
@@ -1,4 +1,4 @@
-package docker
+package changes
 
 import (
 	"fmt"
